Change var type Shouldn't be a float

+++ src/main.c
@@ -18,8 +18,8 @@ int main(void)
 {
     // Initialization
     //--------------------------------------------------------------------------------------
-    const float screen_width = 1000;
-    const float screen_height = 1000;
+    const int screen_width = 1000;
+    const int screen_height = 1000;
 
 
     MemoCache* memo_cache = init_memo_cache();
@@ -56,7 +56,7 @@ int main(void)
         .primary_btn_color = GOLD
     };
 
-    InitWindow((int)screen_width, (int)screen_height, "THE BEST --- Tic Tac Toe");
+    InitWindow(screen_width, screen_height, "THE BEST --- Tic Tac Toe");
     InitAudioDevice();
 
     GameResources resources = load_game_resources();
